requirements.txt
pyaudio
numpy
faiss-cpu
scikit-learn
python-dotenv
openai
websocketscls
